Add SharedStatusLabel component
