+++ src/components/SharedStatusLabel/component.tsx
@@ -1,0 +1,18 @@
+import React, { FunctionComponent } from "react";
+import { StatusLabel } from "$components/StatusLabel";
+import { IComponentProps } from "./interfaces";
+
+export const SharedStatusLabel: FunctionComponent<IComponentProps> = (
+  {
+    translations,
+    status,
+    ...props
+  }
+) => (
+  <StatusLabel
+    text={translations[status]}
+    tooltipText={translations[status]}
+    status={status}
+    {...props}
+  />
+);
