+++ src/components/SharedStatusLabel/container.tsx
@@ -1,11 +1,11 @@
 import React, { Fragment, FunctionComponent } from "react";
 import { useTranslations } from "$hooks/queries";
-import { StatusTitle } from "./component";
+import { SharedStatusLabel } from "./component";
 import { IContainerProps, ITranslations } from "./interfaces";
 
-export const StatusTitleContainer: FunctionComponent<IContainerProps> = props => {
+export const SharedStatusLabelContainer: FunctionComponent<IContainerProps> = props => {
   const translations = useTranslations<ITranslations>("statusLabel");
   if (!translations) return <Fragment />;
 
-  return <StatusTitle translations={translations} {...props} />;
+  return <SharedStatusLabel translations={translations} {...props} />;
 };
