+++ src/components/SharedStatusLabel/index.ts
@@ -1,0 +1,1 @@
+export { SharedStatusLabelContainer as SharedStatusLabel } from "./container";
